Change env variable name for GCP TTS keyfile
This is so we can take advantage of
gerywahyunugraha/heroku-google-application-credentials-buildpack
when deploying to Heroku, which has no persistent storage and so
requires the credentials to be accessible in an env var instead of a
file. The buildpack takes that env and writes it to a file in the slug
at deploy time.